Refactor CosmosDB and MongoVCore Branch Data Providers to extend BaseCachedBranchDataProvider
- Introduced BaseCachedBranchDataProvider as a common base class for caching tree data.
- Updated CosmosDBBranchDataProvider to utilize the new base class, simplifying the implementation and enhancing caching logic.
- Refactored MongoVCoreBranchDataProvider to extend BaseCachedBranchDataProvider, improving code reuse and maintainability.
- Simplified CosmosDBWorkspaceBranchDataProvider and ClustersWorkspaceBranchDataProvider by extending BaseCachedBranchDataProvider, reducing boilerplate code and enhancing clarity.
- Removed redundant caching logic and event emitters in favor of the base class's functionality.
