+++ src/tree/azure-resources-view/documentdb/mongo-vcore/MongoVCoreBranchDataProvider.ts
@@ -6,18 +6,12 @@
 import { type GenericResource } from '@azure/arm-resources';
 import { getResourceGroupFromId, uiUtils } from '@microsoft/vscode-azext-azureutils';
 import { callWithTelemetryAndErrorHandling, nonNullProp, type IActionContext } from '@microsoft/vscode-azext-utils';
-import {
-    type AzureResource,
-    type AzureSubscription,
-    type BranchDataProvider,
-} from '@microsoft/vscode-azureresources-api';
-import * as vscode from 'vscode';
+import { type AzureResource, type AzureSubscription } from '@microsoft/vscode-azureresources-api';
 import { API, MongoClustersExperience } from '../../../../AzureDBExperiences';
-import { ext } from '../../../../extensionVariables';
 import { createMongoClustersManagementClient } from '../../../../utils/azureClients';
+import { BaseCachedBranchDataProvider } from '../../../BaseCachedBranchDataProvider';
 import { type ClusterModel } from '../../../documentdb/ClusterModel';
 import { type TreeElement } from '../../../TreeElement';
-import { isTreeElementWithContextValue } from '../../../TreeElementWithContextValue';
 import { MongoVCoreResourceItem } from './MongoVCoreResourceItem';
 
 export type MongoVCoreResource = AzureResource &
@@ -25,107 +19,61 @@ export type MongoVCoreResource = AzureResource &
         readonly raw: GenericResource; // Resource object from Azure SDK
     };
 
-export class MongoVCoreBranchDataProvider
-    extends vscode.Disposable
-    implements BranchDataProvider<MongoVCoreResource, TreeElement>
-{
+export class MongoVCoreBranchDataProvider extends BaseCachedBranchDataProvider<MongoVCoreResource> {
+    protected get contexValuePrefix(): string {
+        return 'mongoVCore.azure';
+    }
+
     private detailsCacheUpdateRequested = true;
     private detailsCache: Map<string, ClusterModel> = new Map<string, ClusterModel>();
     private itemsToUpdateInfo: Map<string, MongoVCoreResourceItem> = new Map<string, MongoVCoreResourceItem>();
 
-    private readonly onDidChangeTreeDataEmitter = new vscode.EventEmitter<TreeElement | undefined>();
+    protected createResourceItem(context: IActionContext, resource: MongoVCoreResource): TreeElement | undefined {
+        // TODO: ClusterModel does not implement TreeElementWithExperience as other models do
+        // and the base class will check isTreeElementWithExperience to set this property.
+        // Since ClusterModel has a dbExperience property which is theoretically the same but named differently
+        // set the experience to the API.MongoClusters explicitly here.
+        context.telemetry.properties.experience = API.MongoClusters;
 
-    constructor() {
-        super(() => this.onDidChangeTreeDataEmitter.dispose());
-    }
-
-    get onDidChangeTreeData(): vscode.Event<TreeElement | undefined> {
-        return this.onDidChangeTreeDataEmitter.event;
-    }
-
-    async getChildren(element: TreeElement): Promise<TreeElement[]> {
-        /**
-         * getChildren is called for every element in the tree when expanding, the element being expanded is being passed as an argument
-         */
-        const result = await callWithTelemetryAndErrorHandling('getChildren', async (context: IActionContext) => {
-            context.telemetry.properties.experience = API.MongoClusters;
-
-            if (isTreeElementWithContextValue(element)) {
-                context.telemetry.properties.parentNodeContext = element.contextValue;
-            }
-
-            const children = (await element.getChildren?.()) ?? [];
-            return children.map((child) => {
-                return ext.state.wrapItemInStateHandling(child, (child: TreeElement) =>
-                    this.refresh(child),
-                ) as TreeElement;
-            });
-        });
-
-        return result ?? [];
-    }
-
-    async getResourceItem(element: MongoVCoreResource): Promise<TreeElement> {
-        /**
-         * This function is being called when the resource tree is being built, it is called for every resource element in the tree.
-         */
-
-        const resourceItem = await callWithTelemetryAndErrorHandling(
-            'resolveResource',
-            // disabling require-await, the async aspect is in there, but uses the .then pattern
-            // eslint-disable-next-line @typescript-eslint/require-await
-            async (context: IActionContext) => {
-                context.telemetry.properties.experience = API.MongoClusters;
-
-                if (this.detailsCacheUpdateRequested) {
-                    void this.updateResourceCache(context, element.subscription, 1000 * 60 * 5).then(() => {
-                        /**
-                         * Instances of MongoClusterItem were  stored in the itemsToUpdateInfo map,
-                         * so that when the cache is updated, the items can be refreshed.
-                         * I had to keep all of them in the map becasuse refresh requires the actual MongoClusterItem instance.
-                         */
-                        this.itemsToUpdateInfo.forEach((value: MongoVCoreResourceItem) => {
-                            value.cluster = {
-                                ...value.cluster,
-                                ...this.detailsCache.get(value.cluster.id),
-                            };
-                            this.refresh(value);
-                        });
-
-                        this.itemsToUpdateInfo.clear();
-                    });
-                }
-
-                // 1. extract the basic info from the element (subscription, resource group, etc., provided by Azure Resources)
-                let clusterInfo: ClusterModel = {
-                    ...element,
-                    dbExperience: MongoClustersExperience,
-                } as ClusterModel;
-
-                // 2. lookup the details in the cache, on subsequent refreshes, the details will be available in the cache
-                if (this.detailsCache.has(clusterInfo.id)) {
-                    clusterInfo = {
-                        ...clusterInfo,
-                        ...this.detailsCache.get(clusterInfo.id),
+        if (this.detailsCacheUpdateRequested) {
+            void this.updateResourceCache(context, resource.subscription, 1000 * 60 * 5).then(() => {
+                /**
+                 * Instances of MongoClusterItem were  stored in the itemsToUpdateInfo map,
+                 * so that when the cache is updated, the items can be refreshed.
+                 * I had to keep all of them in the map becasuse refresh requires the actual MongoClusterItem instance.
+                 */
+                this.itemsToUpdateInfo.forEach((value: MongoVCoreResourceItem) => {
+                    value.cluster = {
+                        ...value.cluster,
+                        ...this.detailsCache.get(value.cluster.id),
                     };
-                }
+                    this.refresh(value);
+                });
 
-                const clusterItem = new MongoVCoreResourceItem(element.subscription, clusterInfo);
-
-                // 3. store the item in the update queue, so that when the cache is updated, the item can be refreshed
-                this.itemsToUpdateInfo.set(clusterItem.id, clusterItem);
-
-                return clusterItem;
-            },
-        );
-
-        if (resourceItem) {
-            return ext.state.wrapItemInStateHandling(resourceItem, (item: TreeElement) =>
-                this.refresh(item),
-            ) as TreeElement;
+                this.itemsToUpdateInfo.clear();
+            });
         }
 
-        return null as unknown as TreeElement;
+        // 1. extract the basic info from the element (subscription, resource group, etc., provided by Azure Resources)
+        let clusterInfo: ClusterModel = {
+            ...resource,
+            dbExperience: MongoClustersExperience,
+        } as ClusterModel;
+
+        // 2. lookup the details in the cache, on subsequent refreshes, the details will be available in the cache
+        if (this.detailsCache.has(clusterInfo.id)) {
+            clusterInfo = {
+                ...clusterInfo,
+                ...this.detailsCache.get(clusterInfo.id),
+            };
+        }
+
+        const clusterItem = new MongoVCoreResourceItem(resource.subscription, clusterInfo);
+
+        // 3. store the item in the update queue, so that when the cache is updated, the item can be refreshed
+        this.itemsToUpdateInfo.set(clusterItem.id, clusterItem);
+
+        return clusterItem;
     }
 
     async updateResourceCache(
@@ -188,28 +136,5 @@ export class MongoVCoreBranchDataProvider
                 }
             },
         );
-    }
-
-    // onDidChangeTreeData?: vscode.Event<void | TreeElementBase | TreeElementBase[] | null | undefined> | undefined;
-
-    async getTreeItem(element: TreeElement): Promise<vscode.TreeItem> {
-        return element.getTreeItem();
-    }
-
-    /**
-     * Optional method to return the parent of `element`.
-     * Return `null` or `undefined` if `element` is a child of root.
-     *
-     * **NOTE:** This method should be implemented in order to access {@link TreeView.reveal reveal} API.
-     *
-     * @param element The element for which the parent has to be returned.
-     * @returns Parent of `element`.
-     */
-    getParent(element: TreeElement): vscode.ProviderResult<TreeElement> {
-        return element.getParent?.() ?? null;
-    }
-
-    refresh(element?: TreeElement): void {
-        this.onDidChangeTreeDataEmitter.fire(element);
     }
 }
